EES-3222 pr feedback
