+++ src/explore-education-statistics-admin/src/routes/publicationRoutes.ts
@@ -1,6 +1,8 @@
 import PublicationContactPage from '@admin/pages/publication/PublicationContactPage';
 import PublicationDetailsPage from '@admin/pages/publication/PublicationDetailsPage';
-import PublicationMethodologyPage from '@admin/pages/publication/PublicationMethodologyPage';
+import PublicationAdoptMethodologyPage from '@admin/pages/publication/PublicationAdoptMethodologyPage';
+import PublicationExternalMethodologyPage from '@admin/pages/publication/PublicationExternalMethodologyPage';
+import PublicationMethodologiesPage from '@admin/pages/publication/PublicationMethodologiesPage';
 import PublicationReleasesPage from '@admin/pages/publication/PublicationReleasesPage';
 import { RouteProps } from 'react-router';
 
@@ -19,10 +21,22 @@ export const publicationReleasesRoute: PublicationRouteProps = {
   component: PublicationReleasesPage,
 };
 
-export const publicationMethodologyRoute: PublicationRouteProps = {
-  path: '/publication/:publicationId/methodology',
-  title: 'Methodology',
-  component: PublicationMethodologyPage,
+export const publicationMethodologiesRoute: PublicationRouteProps = {
+  path: '/publication/:publicationId/methodologies',
+  title: 'Methodologies',
+  component: PublicationMethodologiesPage,
+};
+
+export const publicationAdoptMethodologyRoute: PublicationRouteProps = {
+  path: '/publication/:publicationId/methodologies/adopt',
+  title: 'Adopt a Methodology',
+  component: PublicationAdoptMethodologyPage,
+};
+
+export const publicationExternalMethodologyRoute: PublicationRouteProps = {
+  path: '/publication/:publicationId/methodologies/external',
+  title: 'External Methodology',
+  component: PublicationExternalMethodologyPage,
 };
 
 export const publicationDetailsRoute: PublicationRouteProps = {
